Tweaks and new model
Added a pier model, renamed a script, rebaked occlusion
