+++ Assets/_project/Prefabs/Pier .prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: bef137a2d9e1fc74fb95015985328265
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
